Extract components for tags
