+++ app/components/show/apo/details_component.html.erb
@@ -11,7 +11,7 @@
       <tr>
         <th class="col-3" scope="row">Project</th>
         <td>
-            <%= project_tag %>
+          <%= render Show::ProjectTagComponent.new(document: @presenter.document) %>
         </td>
       </tr>
 
@@ -25,7 +25,7 @@
       <tr>
         <th class="col-3" scope="row">Tags</th>
         <td>
-            <%= tags %> <%= edit_tags %>
+          <%= render Show::TagsComponent.new(document: @presenter.document) %>
         </td>
       </tr>
     </tbody>
