+++ app/components/show/tags_component.html.erb
@@ -1,0 +1,1 @@
+<%= tags %> <%= edit_tags %>
